Adding files from my computer to this directory as I transition to using GitHub to store and protect my code base.
Files in this directory include:
- a correlation rscript
- a .zsh sript to run the rscript on
a blast result with multiple ASVs of interest in
it.
- A modified version of the rscript for running
correlation on subsetted ASV table which was
filtered based on zone.
- .zsh that eventually was abandoned in favor of
reading everything into the rscript
- a spearman R markdown that runs Spearman correlation
on an ASV table that is filtered on occurrence
and abundance. It outputs the top 30ish significant
results for each ASV that had a 99% identity to
one of the BLO sequences.