I edited helper_asvs_soil.R to sort the input helper_df by descending correlation in the get_helper_abundance funciton.
I also removed zero_sum rows in the asvTable after
reading in the data. The reason zero_sum rows are
present is because the asvTable is a subset of the
full asvTable, and thus some rows that had non-zero_sum
in the original table became zero_sum rows. This was
not handled before.

I also added the relative abundance transformation.

Finally, I output the top 5 samples as a stacked
bar plot of the helper asv abundances.