Create run_blast.sh
run_blast.sh is a script to extract an input FASTA file (with multiple sequences if you so please) and run them against the blast_db you have made. It also filters the results based on an arbitrary percent identity (which you need to edit) and separates each sequence in the file into individual txt files.